Apply refactoring proposed by @jomarko
