+++ kie-wb-common-dmn/kie-wb-common-dmn-client/src/test/java/org/kie/workbench/common/dmn/client/editors/types/listview/draganddrop/DNDDataTypesHandlerTest.java
@@ -20,8 +20,6 @@ import java.util.Optional;
 
 import com.google.gwtmockito.GwtMockitoTestRunner;
 import elemental2.dom.Element;
-import elemental2.dom.HTMLElement;
-import elemental2.dom.NodeList;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -29,28 +27,24 @@ import org.kie.workbench.common.dmn.client.editors.types.common.DataType;
 import org.kie.workbench.common.dmn.client.editors.types.common.DataTypeManager;
 import org.kie.workbench.common.dmn.client.editors.types.listview.DataTypeList;
 import org.kie.workbench.common.dmn.client.editors.types.listview.DataTypeListItem;
-import org.kie.workbench.common.dmn.client.editors.types.listview.draganddrop.DNDDataTypesHandler.DNDContext;
-import org.kie.workbench.common.dmn.client.editors.types.listview.draganddrop.DNDDataTypesHandler.ShiftStrategy;
 import org.kie.workbench.common.dmn.client.editors.types.persistence.DataTypeStore;
 import org.kie.workbench.common.dmn.client.editors.types.persistence.ItemDefinitionStore;
 import org.mockito.Mock;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static org.kie.workbench.common.dmn.client.editors.types.listview.DataTypeListItemView.UUID_ATTR;
-import static org.kie.workbench.common.dmn.client.editors.types.listview.draganddrop.DNDDataTypesHandler.ShiftStrategy.INSERT_INTO_HOVERED_DATA_TYPE;
-import static org.kie.workbench.common.dmn.client.editors.types.listview.draganddrop.DNDDataTypesHandler.ShiftStrategy.INSERT_NESTED_DATA_TYPE;
-import static org.kie.workbench.common.dmn.client.editors.types.listview.draganddrop.DNDDataTypesHandler.ShiftStrategy.INSERT_SIBLING_DATA_TYPE;
-import static org.kie.workbench.common.dmn.client.editors.types.listview.draganddrop.DNDDataTypesHandler.ShiftStrategy.INSERT_TOP_LEVEL_DATA_TYPE;
-import static org.kie.workbench.common.dmn.client.editors.types.listview.draganddrop.DNDDataTypesHandler.ShiftStrategy.INSERT_TOP_LEVEL_DATA_TYPE_AT_THE_TOP;
-import static org.kie.workbench.common.dmn.client.editors.types.listview.draganddrop.DNDListDOMHelper.DATA_X_POSITION;
-import static org.kie.workbench.common.dmn.client.editors.types.listview.draganddrop.DNDListDOMHelper.DATA_Y_POSITION;
+import static org.kie.workbench.common.dmn.client.editors.types.listview.draganddrop.DNDDataTypesHandlerShiftStrategy.INSERT_INTO_HOVERED_DATA_TYPE;
+import static org.kie.workbench.common.dmn.client.editors.types.listview.draganddrop.DNDDataTypesHandlerShiftStrategy.INSERT_NESTED_DATA_TYPE;
+import static org.kie.workbench.common.dmn.client.editors.types.listview.draganddrop.DNDDataTypesHandlerShiftStrategy.INSERT_SIBLING_DATA_TYPE;
+import static org.kie.workbench.common.dmn.client.editors.types.listview.draganddrop.DNDDataTypesHandlerShiftStrategy.INSERT_TOP_LEVEL_DATA_TYPE;
+import static org.kie.workbench.common.dmn.client.editors.types.listview.draganddrop.DNDDataTypesHandlerShiftStrategy.INSERT_TOP_LEVEL_DATA_TYPE_AT_THE_TOP;
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.spy;
@@ -92,8 +86,8 @@ public class DNDDataTypesHandlerTest {
         final Element hoverElement = mock(Element.class);
         final DataType current = mock(DataType.class);
         final DataType reference = mock(DataType.class);
-        final DNDContext context = mock(DNDContext.class);
-        final ShiftStrategy strategy = INSERT_INTO_HOVERED_DATA_TYPE;
+        final DNDDataTypesHandlerContext context = mock(DNDDataTypesHandlerContext.class);
+        final DNDDataTypesHandlerShiftStrategy strategy = INSERT_INTO_HOVERED_DATA_TYPE;
 
         doNothing().when(handler).logError(anyString());
         doReturn(context).when(handler).makeDndContext(currentElement, hoverElement);
@@ -112,7 +106,7 @@ public class DNDDataTypesHandlerTest {
         final Element currentElement = mock(Element.class);
         final Element hoverElement = mock(Element.class);
         final DataType reference = mock(DataType.class);
-        final DNDContext context = mock(DNDContext.class);
+        final DNDDataTypesHandlerContext context = mock(DNDDataTypesHandlerContext.class);
 
         doNothing().when(handler).logError(anyString());
         doReturn(context).when(handler).makeDndContext(currentElement, hoverElement);
@@ -130,7 +124,7 @@ public class DNDDataTypesHandlerTest {
         final Element currentElement = mock(Element.class);
         final Element hoverElement = mock(Element.class);
         final DataType current = mock(DataType.class);
-        final DNDContext context = mock(DNDContext.class);
+        final DNDDataTypesHandlerContext context = mock(DNDDataTypesHandlerContext.class);
 
         doNothing().when(handler).logError(anyString());
         doReturn(context).when(handler).makeDndContext(currentElement, hoverElement);
@@ -149,10 +143,11 @@ public class DNDDataTypesHandlerTest {
         final Element hoverElement = mock(Element.class);
 
         doNothing().when(handler).logError(anyString());
+        doThrow(new UnsupportedOperationException("Error")).when(handler).makeDndContext(any(), any());
 
         handler.onDropDataType(currentElement, hoverElement);
 
-        verify(handler).logError("Drag-n-Drop error. Check 'DNDDataTypesHandler'.");
+        verify(handler).logError("Drag-n-Drop error (Error). Check 'DNDDataTypesHandler'.");
     }
 
     @Test
@@ -161,7 +156,7 @@ public class DNDDataTypesHandlerTest {
         final DataType current = mock(DataType.class);
         final DataType clone = mock(DataType.class);
         final DataType reference = mock(DataType.class);
-        final ShiftStrategy strategy = INSERT_INTO_HOVERED_DATA_TYPE;
+        final DNDDataTypesHandlerShiftStrategy strategy = INSERT_INTO_HOVERED_DATA_TYPE;
         final String referenceHash = "referenceHash";
         final DataTypeListItem oldItem = mock(DataTypeListItem.class);
         final DataTypeListItem referenceItem = mock(DataTypeListItem.class);
@@ -188,7 +183,7 @@ public class DNDDataTypesHandlerTest {
         final DataType current = mock(DataType.class);
         final DataType clone = mock(DataType.class);
         final DataType reference = mock(DataType.class);
-        final ShiftStrategy strategy = INSERT_INTO_HOVERED_DATA_TYPE;
+        final DNDDataTypesHandlerShiftStrategy strategy = INSERT_INTO_HOVERED_DATA_TYPE;
         final String referenceHash = "referenceHash";
         final DataTypeListItem oldItem = mock(DataTypeListItem.class);
         final DataTypeListItem referenceItem = mock(DataTypeListItem.class);
@@ -210,309 +205,10 @@ public class DNDDataTypesHandlerTest {
     }
 
     @Test
-    public void testDNDContextGetReferenceWhenHoveredDataTypeIsPresent() {
-
-        final Element currentElement = mock(Element.class);
-        final Element hoverElement = mock(Element.class);
-        final Optional<DataType> hoverDataType = Optional.of(mock(DataType.class));
-        final String uuid = "0000-1111-2222-3333";
-
-        when(hoverElement.getAttribute(UUID_ATTR)).thenReturn(uuid);
-        when(dataTypeStore.get(uuid)).thenReturn(hoverDataType.get());
-
-        final Optional<DataType> reference = handler.makeDndContext(currentElement, hoverElement).getReference();
-
-        assertEquals(hoverDataType, reference);
-    }
-
-    @Test
-    public void testDNDContextGetReferenceWhenHoveredDataTypeIsPresentButHoveredDataTypeIsReadOnly() {
-
-        final Element currentElement = mock(Element.class);
-        final Element hoverElement = mock(Element.class);
-        final Element previousElement = mock(Element.class);
-        final Optional<DataType> hoverDataType = Optional.of(mock(DataType.class));
-        final Optional<DataType> previousDataType = Optional.of(mock(DataType.class));
-        final String hoverUUID = "0000-0000-0000-0000";
-        final String previousUUID = "1111-1111-1111-1111";
-
-        when(dndListComponent.getPreviousElement(any(), any())).thenReturn(Optional.of(previousElement));
-        when(dataTypeStore.get(previousUUID)).thenReturn(previousDataType.get());
-        when(dataTypeStore.get(hoverUUID)).thenReturn(hoverDataType.get());
-        when(previousElement.getAttribute(UUID_ATTR)).thenReturn(previousUUID);
-        when(hoverElement.getAttribute(UUID_ATTR)).thenReturn(hoverUUID);
-        when(hoverDataType.get().isReadOnly()).thenReturn(true);
-
-        final Optional<DataType> reference = handler.makeDndContext(currentElement, hoverElement).getReference();
-
-        assertEquals(previousDataType, reference);
-    }
-
-    @Test
-    public void testDNDContextGetReferenceWhenPreviousDataTypeIsPresent() {
-
-        final Element currentElement = mock(Element.class);
-        final Element hoverElement = mock(Element.class);
-        final Element previousElement = mock(Element.class);
-        final Optional<DataType> previousDataType = Optional.of(mock(DataType.class));
-        final String uuid = "0000-1111-2222-3333";
-
-        when(dndListComponent.getPreviousElement(any(), any())).thenReturn(Optional.of(previousElement));
-        when(dataTypeStore.get(uuid)).thenReturn(previousDataType.get());
-        when(previousElement.getAttribute(UUID_ATTR)).thenReturn(uuid);
-
-        final Optional<DataType> reference = handler.makeDndContext(currentElement, hoverElement).getReference();
-
-        assertEquals(previousDataType, reference);
-    }
-
-    @Test
-    public void testDNDContextGetReferenceWhenCurrentDataTypeIsPresent() {
-
-        final Element currentElement = mock(Element.class);
-        final Element hoverElement = mock(Element.class);
-        final Element firstElement = mock(Element.class);
-        final Element hiddenElement = mock(Element.class);
-        final HTMLElement dragArea = mock(HTMLElement.class);
-        final Optional<DataType> currentDataType = Optional.of(mock(DataType.class));
-        final Optional<DataType> firstDataType = Optional.of(mock(DataType.class));
-        final String currentUUID = "0000-0000-0000-0000";
-        final String firstUUID = "1111-1111-1111-1111";
-
-        dragArea.childNodes = spy(new NodeList<>());
-        dragArea.childNodes.length = 3;
-
-        doReturn(hiddenElement).when(dragArea.childNodes).getAt(0);
-        doReturn(currentElement).when(dragArea.childNodes).getAt(1);
-        doReturn(firstElement).when(dragArea.childNodes).getAt(2);
-
-        when(hiddenElement.getAttribute(DATA_Y_POSITION)).thenReturn("-1");
-        when(currentElement.getAttribute(DATA_Y_POSITION)).thenReturn("0");
-        when(firstElement.getAttribute(DATA_Y_POSITION)).thenReturn("1");
-        when(hiddenElement.getAttribute(DATA_X_POSITION)).thenReturn("0");
-        when(currentElement.getAttribute(DATA_X_POSITION)).thenReturn("0");
-        when(firstElement.getAttribute(DATA_X_POSITION)).thenReturn("0");
-        when(dndListComponent.getDragArea()).thenReturn(dragArea);
-        when(dndListComponent.getPreviousElement(any(), any())).thenReturn(Optional.empty());
-        when(dataTypeStore.get(currentUUID)).thenReturn(currentDataType.get());
-        when(dataTypeStore.get(firstUUID)).thenReturn(firstDataType.get());
-        when(currentDataType.get().getName()).thenReturn("Current Data Type");
-        when(firstDataType.get().getName()).thenReturn("First Data Type");
-        when(currentElement.getAttribute(UUID_ATTR)).thenReturn(currentUUID);
-        when(firstElement.getAttribute(UUID_ATTR)).thenReturn(firstUUID);
-
-        final Optional<DataType> reference = handler.makeDndContext(currentElement, hoverElement).getReference();
-
-        assertEquals(firstDataType, reference);
-    }
-
-    @Test
-    public void testDNDContextGetReferenceWhenDataTypeListIsNotInitialized() {
-
-        final Element currentElement = mock(Element.class);
-        final Element hoverElement = mock(Element.class);
-        final Element previousElement = mock(Element.class);
-        final Optional<DataType> previousDataType = Optional.of(mock(DataType.class));
-        final String uuid = "0000-1111-2222-3333";
-
-        when(dndListComponent.getPreviousElement(any(), any())).thenReturn(Optional.of(previousElement));
-        when(dataTypeStore.get(uuid)).thenReturn(previousDataType.get());
-        when(previousElement.getAttribute(UUID_ATTR)).thenReturn(uuid);
-
-        handler.init(null);
-
-        assertThatThrownBy(() -> handler.makeDndContext(currentElement, hoverElement).getReference())
-                .isInstanceOf(UnsupportedOperationException.class)
-                .hasMessage("'DNDDataTypesHandler' must be initialized with a 'DataTypeList' instance.");
-    }
-
-    @Test
-    public void testDNDContextGetReferenceWhenDNDListComponentIsNotInitialized() {
-
-        final Element currentElement = mock(Element.class);
-        final Element hoverElement = mock(Element.class);
-        final Element previousElement = mock(Element.class);
-        final Optional<DataType> previousDataType = Optional.of(mock(DataType.class));
-        final String uuid = "0000-1111-2222-3333";
-
-        when(dndListComponent.getPreviousElement(any(), any())).thenReturn(Optional.of(previousElement));
-        when(dataTypeStore.get(uuid)).thenReturn(previousDataType.get());
-        when(previousElement.getAttribute(UUID_ATTR)).thenReturn(uuid);
-
-        when(dataTypeList.getDNDListComponent()).thenReturn(null);
-
-        assertThatThrownBy(() -> handler.makeDndContext(currentElement, hoverElement).getReference())
-                .isInstanceOf(UnsupportedOperationException.class)
-                .hasMessage("'DNDDataTypesHandler' must be initialized with a 'DNDListComponent' instance.");
-    }
-
-    @Test
-    public void testDNDContextGetReferenceWhenItIsNotPresent() {
-
-        final Element currentElement = mock(Element.class);
-        final Element hoverElement = mock(Element.class);
-
-        when(dndListComponent.getPreviousElement(any(), any())).thenReturn(Optional.empty());
-
-        final Optional<DataType> reference = handler.makeDndContext(currentElement, hoverElement).getReference();
-
-        assertFalse(reference.isPresent());
-    }
-
-    @Test
-    public void testGetStrategyInsertIntoHoveredDataType() {
-
-        final Element currentElement = mock(Element.class);
-        final Element hoverElement = mock(Element.class);
-        final DNDContext context = handler.makeDndContext(currentElement, hoverElement);
-        final Optional<DataType> hoverDataType = Optional.of(mock(DataType.class));
-        final String uuid = "0000-1111-2222-3333";
-
-        when(hoverElement.getAttribute(UUID_ATTR)).thenReturn(uuid);
-        when(dataTypeStore.get(uuid)).thenReturn(hoverDataType.get());
-        loadReferenceContext(context);
-
-        final ShiftStrategy actualShiftStrategy = context.getStrategy();
-        final ShiftStrategy expectedShiftStrategy = INSERT_INTO_HOVERED_DATA_TYPE;
-
-        assertEquals(expectedShiftStrategy, actualShiftStrategy);
-    }
-
-    @Test
-    public void testGetStrategyInsertIntoHoveredDataTypeWhenHoveredDataTypeIsReadOnly() {
-
-        final Element currentElement = mock(Element.class);
-        final Element hoverElement = mock(Element.class);
-        final DNDContext context = handler.makeDndContext(currentElement, hoverElement);
-        final Optional<DataType> hoverDataType = Optional.of(mock(DataType.class));
-        final String uuid = "0000-1111-2222-3333";
-
-        when(hoverDataType.get().isReadOnly()).thenReturn(true);
-        when(hoverElement.getAttribute(UUID_ATTR)).thenReturn(uuid);
-        when(dataTypeStore.get(uuid)).thenReturn(hoverDataType.get());
-        when(dndListComponent.getPreviousElement(any(), any())).thenReturn(Optional.empty());
-        loadReferenceContext(context);
-
-        final ShiftStrategy actualShiftStrategy = context.getStrategy();
-        final ShiftStrategy expectedShiftStrategy = INSERT_TOP_LEVEL_DATA_TYPE_AT_THE_TOP;
-
-        assertEquals(expectedShiftStrategy, actualShiftStrategy);
-    }
-
-    @Test
-    public void testGetStrategyInsertTopLevelDataTypeAtTheTop() {
-
-        final Element currentElement = mock(Element.class);
-        final Element hoverElement = mock(Element.class);
-        final DNDContext context = handler.makeDndContext(currentElement, hoverElement);
-
-        when(dndListComponent.getPreviousElement(any(), any())).thenReturn(Optional.empty());
-        loadReferenceContext(context);
-
-        final ShiftStrategy actualShiftStrategy = context.getStrategy();
-        final ShiftStrategy expectedShiftStrategy = INSERT_TOP_LEVEL_DATA_TYPE_AT_THE_TOP;
-
-        assertEquals(expectedShiftStrategy, actualShiftStrategy);
-    }
-
-    @Test
-    public void testGetStrategyInsertTopLevelDataType() {
-
-        final Element currentElement = mock(Element.class);
-        final Element hoverElement = mock(Element.class);
-        final Element previousElement = mock(Element.class);
-        final DNDContext context = handler.makeDndContext(currentElement, hoverElement);
-        final String uuid = "0000-1111-2222-3333";
-        final Optional<DataType> previousDataType = Optional.of(mock(DataType.class));
-
-        when(dataTypeStore.get(uuid)).thenReturn(previousDataType.get());
-        when(dndListComponent.getPreviousElement(any(), any())).thenReturn(Optional.of(previousElement));
-        when(currentElement.getAttribute(DATA_X_POSITION)).thenReturn("0");
-        when(previousElement.getAttribute(DATA_X_POSITION)).thenReturn("0");
-        when(previousElement.getAttribute(UUID_ATTR)).thenReturn(uuid);
-        loadReferenceContext(context);
-
-        final ShiftStrategy actualShiftStrategy = context.getStrategy();
-        final ShiftStrategy expectedShiftStrategy = INSERT_TOP_LEVEL_DATA_TYPE;
-
-        assertEquals(expectedShiftStrategy, actualShiftStrategy);
-    }
-
-    @Test
-    public void testGetStrategyInsertNestedDataType() {
-
-        final Element currentElement = mock(Element.class);
-        final Element hoverElement = mock(Element.class);
-        final Element previousElement = mock(Element.class);
-        final DNDContext context = handler.makeDndContext(currentElement, hoverElement);
-        final String uuid = "0000-1111-2222-3333";
-        final Optional<DataType> previousDataType = Optional.of(mock(DataType.class));
-
-        when(dataTypeStore.get(uuid)).thenReturn(previousDataType.get());
-        when(dndListComponent.getPreviousElement(any(), any())).thenReturn(Optional.of(previousElement));
-        when(currentElement.getAttribute(DATA_X_POSITION)).thenReturn("1");
-        when(previousElement.getAttribute(DATA_X_POSITION)).thenReturn("0");
-        when(previousElement.getAttribute(UUID_ATTR)).thenReturn(uuid);
-        loadReferenceContext(context);
-
-        final ShiftStrategy actualShiftStrategy = context.getStrategy();
-        final ShiftStrategy expectedShiftStrategy = INSERT_NESTED_DATA_TYPE;
-
-        assertEquals(expectedShiftStrategy, actualShiftStrategy);
-    }
-
-    @Test
-    public void testGetStrategyInsertNestedDataTypeWhenPreviousDataTypeIsReadOnly() {
-
-        final Element currentElement = mock(Element.class);
-        final Element hoverElement = mock(Element.class);
-        final Element previousElement = mock(Element.class);
-        final DNDContext context = handler.makeDndContext(currentElement, hoverElement);
-        final String uuid = "0000-1111-2222-3333";
-        final Optional<DataType> previousDataType = Optional.of(mock(DataType.class));
-
-        when(previousDataType.get().isReadOnly()).thenReturn(true);
-        when(dataTypeStore.get(uuid)).thenReturn(previousDataType.get());
-        when(dndListComponent.getPreviousElement(any(), any())).thenReturn(Optional.of(previousElement));
-        when(currentElement.getAttribute(DATA_X_POSITION)).thenReturn("1");
-        when(previousElement.getAttribute(DATA_X_POSITION)).thenReturn("0");
-        when(previousElement.getAttribute(UUID_ATTR)).thenReturn(uuid);
-        loadReferenceContext(context);
-
-        final ShiftStrategy actualShiftStrategy = context.getStrategy();
-        final ShiftStrategy expectedShiftStrategy = INSERT_SIBLING_DATA_TYPE;
-
-        assertEquals(expectedShiftStrategy, actualShiftStrategy);
-    }
-
-    @Test
-    public void testGetStrategyInsertSiblingDataType() {
-
-        final Element currentElement = mock(Element.class);
-        final Element hoverElement = mock(Element.class);
-        final Element previousElement = mock(Element.class);
-        final DNDContext context = handler.makeDndContext(currentElement, hoverElement);
-        final String uuid = "0000-1111-2222-3333";
-        final Optional<DataType> previousDataType = Optional.of(mock(DataType.class));
-
-        when(dataTypeStore.get(uuid)).thenReturn(previousDataType.get());
-        when(dndListComponent.getPreviousElement(any(), any())).thenReturn(Optional.of(previousElement));
-        when(currentElement.getAttribute(DATA_X_POSITION)).thenReturn("1");
-        when(previousElement.getAttribute(DATA_X_POSITION)).thenReturn("1");
-        when(previousElement.getAttribute(UUID_ATTR)).thenReturn(uuid);
-        loadReferenceContext(context);
-
-        final ShiftStrategy actualShiftStrategy = context.getStrategy();
-        final ShiftStrategy expectedShiftStrategy = INSERT_SIBLING_DATA_TYPE;
-
-        assertEquals(expectedShiftStrategy, actualShiftStrategy);
-    }
-
-    @Test
     public void testIsTopLevelShiftOperationWhenDataTypeIsNotTopLevel() {
 
         final DataType dataType = mock(DataType.class);
-        final ShiftStrategy shiftStrategy = INSERT_TOP_LEVEL_DATA_TYPE_AT_THE_TOP;
+        final DNDDataTypesHandlerShiftStrategy shiftStrategy = INSERT_TOP_LEVEL_DATA_TYPE_AT_THE_TOP;
 
         when(dataType.isTopLevel()).thenReturn(false);
 
@@ -542,7 +238,43 @@ public class DNDDataTypesHandlerTest {
         assertTrue(handler.isTopLevelShiftOperation(dataType, INSERT_TOP_LEVEL_DATA_TYPE));
     }
 
-    private void loadReferenceContext(final DNDContext context) {
-        context.getReference();
+    @Test
+    public void testGetReferenceWhenDataTypeListIsNotInitialized() {
+
+        final Element currentElement = mock(Element.class);
+        final Element hoverElement = mock(Element.class);
+        final Element previousElement = mock(Element.class);
+        final Optional<DataType> previousDataType = Optional.of(mock(DataType.class));
+        final String uuid = "0000-1111-2222-3333";
+
+        when(dndListComponent.getPreviousElement(any(), any())).thenReturn(Optional.of(previousElement));
+        when(dataTypeStore.get(uuid)).thenReturn(previousDataType.get());
+        when(previousElement.getAttribute(UUID_ATTR)).thenReturn(uuid);
+
+        handler.init(null);
+
+        assertThatThrownBy(() -> handler.makeDndContext(currentElement, hoverElement).getReference())
+                .isInstanceOf(UnsupportedOperationException.class)
+                .hasMessage("'DNDDataTypesHandler' must be initialized with a 'DataTypeList' instance.");
+    }
+
+    @Test
+    public void testGetReferenceWhenDNDListComponentIsNotInitialized() {
+
+        final Element currentElement = mock(Element.class);
+        final Element hoverElement = mock(Element.class);
+        final Element previousElement = mock(Element.class);
+        final Optional<DataType> previousDataType = Optional.of(mock(DataType.class));
+        final String uuid = "0000-1111-2222-3333";
+
+        when(dndListComponent.getPreviousElement(any(), any())).thenReturn(Optional.of(previousElement));
+        when(dataTypeStore.get(uuid)).thenReturn(previousDataType.get());
+        when(previousElement.getAttribute(UUID_ATTR)).thenReturn(uuid);
+
+        when(dataTypeList.getDNDListComponent()).thenReturn(null);
+
+        assertThatThrownBy(() -> handler.makeDndContext(currentElement, hoverElement).getReference())
+                .isInstanceOf(UnsupportedOperationException.class)
+                .hasMessage("'DNDDataTypesHandler' must be initialized with a 'DNDListComponent' instance.");
     }
 }
